refactor: extract b2bclient
