+++ core/client/b2b-client.ts
@@ -1,0 +1,59 @@
+import { z } from 'zod';
+
+interface LoginWithB2BParams {
+  customerId: number;
+  customerAccessToken: {
+    value: string;
+    expiresAt: string;
+  };
+}
+
+type ENV_VARIABLES = 'B2B_API_TOKEN' | 'BIGCOMMERCE_CHANNEL_ID';
+
+const getEnv = (key: ENV_VARIABLES) => {
+  if (!process.env[key]) {
+    throw new Error(`Environment variable ${key} is not set`);
+  }
+
+  return process.env[key];
+};
+
+async function login({ customerId, customerAccessToken }: LoginWithB2BParams) {
+  const channelId = getEnv('BIGCOMMERCE_CHANNEL_ID');
+
+  const payload = {
+    channelId,
+    customerId,
+    customerAccessToken,
+  };
+
+  const response = await fetch('https://api-b2b.bigcommerce.com/api/io/auth/customers/storefront', {
+    method: 'POST',
+    headers: {
+      Accept: 'application/json',
+      'Content-Type': 'application/json',
+      authToken: getEnv('B2B_API_TOKEN'),
+    },
+    body: JSON.stringify(payload),
+  });
+
+  const B2BTokenResponseSchema = z.object({
+    data: z.object({
+      token: z.array(z.string()),
+    }),
+  });
+
+  const {
+    data: { token },
+  } = B2BTokenResponseSchema.parse(await response.json());
+
+  if (!token[0]) {
+    throw new Error('No token returned from B2B API');
+  }
+
+  return token[0];
+}
+
+export const b2bClient = {
+  login,
+};
